+++ test/bacnet/basic/object/lo/src/main.c
@@ -24,7 +24,7 @@
  */
 static bool is_float_equal(float x1, float x2)
 {
-    return fabs(x1 - x2) < 0.001;
+    return fabsf(x1 - x2) < 0.001f;
 }
 
 /**
